go tests compose service and script
