+++ go-pubsub/order/go.mod
@@ -1,6 +1,6 @@
-module main
+module order
 
-go 1.21.5
+go 1.21.6
 
 require github.com/gin-gonic/gin v1.9.1
 
@@ -13,7 +13,6 @@ require (
 	github.com/go-playground/universal-translator v0.18.1 // indirect
 	github.com/go-playground/validator/v10 v10.14.0 // indirect
 	github.com/goccy/go-json v0.10.2 // indirect
-	github.com/google/go-cmp v0.5.9 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
 	github.com/klauspost/cpuid/v2 v2.2.4 // indirect
 	github.com/leodido/go-urn v1.2.4 // indirect
